Collapse more of the metadata parsing code for all file versions.
